# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HogeDD API
  version: 1.0.0
servers:
  - url: https://api.hogedd.com
paths:
  /health:
    get:
      summary: 稼働確認
      responses:
        "200":
          description: リクエストを処理可能
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    const: ok
    head:
      summary: 稼働確認（本文なし）
      responses:
        "200":
          description: リクエストを処理可能
  /v1/apps:
    get:
      summary: 公開済みアプリ一覧
      responses:
        "200":
          description: 公開済みアプリを配列で返す
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublishedApp"
        "500":
          $ref: "#/components/responses/InternalError"
    head:
      summary: 公開済みアプリ一覧の状態確認（本文なし）
      responses:
        "200":
          description: 取得可能
        "500":
          description: サーバーエラー
  /v1/apps/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
        description: アプリを識別する小文字英数字とハイフンの文字列
    get:
      summary: 公開済みアプリを1件取得
      responses:
        "200":
          description: 公開済みアプリ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishedApp"
        "404":
          $ref: "#/components/responses/AppNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    head:
      summary: 公開済みアプリの状態確認（本文なし）
      responses:
        "200":
          description: 取得可能
        "404":
          description: 存在しないか未公開
        "500":
          description: サーバーエラー
  /v1/me:
    get:
      summary: 現在の認証主体を取得
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Auth0で検証済みの認証主体
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedIdentity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    head:
      summary: 現在の認証主体を取得可能か確認（本文なし）
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 取得可能
        "401":
          description: Access Tokenがないか無効
  /v1/users/me:
    get:
      summary: 現在のHogeDD Userを取得
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 認証主体に紐づく登録済みUser
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: 認証主体に紐づくUserが未登録
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          $ref: "#/components/responses/InternalError"
    head:
      summary: 現在のHogeDD Userを取得可能か確認
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 取得可能
        "401":
          description: Access Tokenがないか無効
        "404":
          description: 認証主体に紐づくUserが未登録
    put:
      summary: 現在の認証主体をHogeDD Userとして登録
      description: 初回はmemberとして作成し、登録済みの場合はroleとstatusを維持してemail snapshotを更新する。
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 登録済みUserの連絡先snapshotを更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "201":
          description: Userを新規登録
          headers:
            Location:
              schema:
                type: string
                const: /v1/users/me
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: request bodyが指定されている
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "405":
          description: GET・HEAD・PUT以外のHTTP method
          headers:
            Allow:
              schema:
                type: string
                const: GET, HEAD, PUT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          description: Auth0 user profileを取得できない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /v1/users/me/profile:
    get:
      summary: 現在のHogeDD Userのプロフィールを取得
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 登録済みプロフィール
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Userが無効
        "404":
          description: Userまたはプロフィールが未登録
    head:
      summary: 現在のプロフィールを取得可能か確認
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 取得可能
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Userが無効
        "404":
          description: Userまたはプロフィールが未登録
    put:
      summary: 現在のHogeDD Userのプロフィールを登録・更新
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [display_name]
              properties:
                display_name:
                  type: string
                  minLength: 1
                  maxLength: 50
      responses:
        "200":
          description: 保存したプロフィール
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: JSONが不正
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Userが無効
        "404":
          description: Userが未登録
        "422":
          description: 表示名が不正
  /v1/management/me:
    get:
      summary: 現在の運営Userを取得
      description: activeなowner・adminだけを返す。認証・認可の失敗理由は404へ秘匿する。
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 認可済みの運営User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementUser"
        "404":
          $ref: "#/components/responses/ManagementNotFound"
    head:
      summary: 運営Userとして利用可能か確認
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 利用可能
        "404":
          description: route、認証、認可、User状態のいずれかを理由に利用不可
  /v1/management/apps:
    get:
      summary: 運営用App一覧を取得
      description: draftを含む全Appを返す。activeなowner・adminだけが利用できる。
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 運営用App一覧
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ManagementApp"
        "404":
          $ref: "#/components/responses/ManagementNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/management/apps/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
    get:
      summary: 運営用App詳細を取得
      security:
        - bearerAuth: []
      responses:
        "200":
          description: versionを含む運営用App詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementApp"
        "404":
          $ref: "#/components/responses/ManagementNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/management/apps/{slug}/publication:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
    put:
      summary: 公開準備中Appを公開
      description: 公開日時はサーバーが確定する。公開済みAppへの再実行は現在値を返す。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [development_drive, youtube_url, version]
              properties:
                development_drive:
                  type: string
                  minLength: 1
                youtube_url:
                  type: string
                  format: uri
                version:
                  type: integer
                  format: int64
                  minimum: 1
      responses:
        "200":
          description: 公開済みApp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementApp"
        "400":
          description: JSON形式またはrequest bodyが不正
        "404":
          $ref: "#/components/responses/ManagementNotFound"
        "409":
          description: version競合
        "422":
          description: 公開条件がdomain ruleを満たさない
        "500":
          $ref: "#/components/responses/InternalError"
    head:
      summary: 運営用App詳細を取得可能か確認
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 取得可能
        "404":
          $ref: "#/components/responses/ManagementNotFound"
    put:
      summary: 公開準備中Appの表示情報を更新
      description: 取得時のversionが現在値と一致する場合だけ更新する。
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateManagementApp"
      responses:
        "200":
          description: 更新後の運営用App
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementApp"
        "400":
          description: JSON形式またはrequest bodyが不正
        "404":
          $ref: "#/components/responses/ManagementNotFound"
        "409":
          description: version競合または公開済みApp
        "422":
          description: App入力がdomain ruleを満たさない
        "500":
          $ref: "#/components/responses/InternalError"
    head:
      summary: 運営用App一覧を取得可能か確認
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 取得可能
        "404":
          $ref: "#/components/responses/ManagementNotFound"
    post:
      summary: 公開準備中Appを作成
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateManagementApp"
      responses:
        "201":
          description: 作成した公開準備中App
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagementApp"
        "400":
          description: JSON形式またはrequest bodyが不正
        "404":
          $ref: "#/components/responses/ManagementNotFound"
        "409":
          description: 同じslugのAppが存在する
        "422":
          description: App入力がdomain ruleを満たさない
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthenticatedIdentity:
      type: object
      required: [issuer, subject]
      properties:
        issuer:
          type: string
          format: uri
        subject:
          type: string
    User:
      type: object
      required: [id, email, email_verified, role, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        email_verified:
          type: boolean
        role:
          type: string
          enum: [owner, admin, member]
        status:
          type: string
          enum: [active, disabled]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserProfile:
      type: object
      additionalProperties: false
      required: [display_name]
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 50
    ManagementUser:
      type: object
      additionalProperties: false
      required: [id, role]
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, admin]
    ManagementApp:
      type: object
      additionalProperties: false
      required: [slug, title, description, status, tags]
      properties:
        slug:
          type: string
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          minLength: 1
          maxLength: 1000
        status:
          type: string
          enum: [preparing, published]
        tags:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 30
        version:
          type: integer
          format: int64
          minimum: 1
        published_at:
          type: string
          format: date-time
        development_drive:
          type: string
        youtube_url:
          type: string
          format: uri
    CreateManagementApp:
      type: object
      additionalProperties: false
      required: [slug, title, description, tags]
      properties:
        slug:
          type: string
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          minLength: 1
          maxLength: 1000
        tags:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 30
    UpdateManagementApp:
      type: object
      additionalProperties: false
      required: [title, description, tags, version]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          minLength: 1
          maxLength: 1000
        tags:
          type: array
          maxItems: 10
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 30
        version:
          type: integer
          format: int64
          minimum: 1
    PublishedApp:
      type: object
      required: [slug, title, description, status, published_at, tags, development_drive, youtube_url]
      properties:
        slug:
          type: string
          example: clean-tasks
        title:
          type: string
          example: Clean Tasks
        description:
          type: string
        status:
          type: string
          const: published
        published_at:
          type: string
          format: date-time
          example: 2026-05-30T00:00:00Z
        tags:
          type: array
          items:
            type: string
        development_drive:
          type: string
        youtube_url:
          type: string
          format: uri
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
  responses:
    Unauthorized:
      description: Access Tokenがないか無効
      headers:
        WWW-Authenticate:
          schema:
            type: string
            const: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    AppNotFound:
      description: 存在しないか未公開
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    InternalError:
      description: サーバーエラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    ManagementNotFound:
      description: routeが存在しないか、運営境界へアクセスできない
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"